% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover.R
\name{discover}
\alias{discover}
\title{graph convolution-based functional co-response group discovery
https://doi.org/10.1109/TCBBIO.2025.3560853}
\usage{
discover(
  M,
  y = NULL,
  pk = NULL,
  fitf = "nmax",
  ytype = "continuous",
  funcS = "+",
  Nmax = NULL,
  alpha = 40,
  maxIter = 500,
  maxFitness = 1,
  run = 100,
  popSize = 200,
  parallel = TRUE,
  monitor = FALSE
)
}
\arguments{
\item{M}{Numeric matrix (samples × taxa). values are topological abundance.}

\item{y}{Functional variable. Can be continuous, binary or categrial depending on \code{ytype}.}

\item{pk}{Optional vector of prior knowledge (same length as ncol(M)).
Elements equal to 1 indicate mandatory inclusion, 0 otherwise.}

\item{fitf}{Fitness function type. Options: \code{"nmax"} or \code{"regul"}.}

\item{ytype}{Type of functional variable. Options: \code{"continuous"} or \code{"binary"} or \code{"categrial"}.}

\item{funcS}{Direction of association for binary outcome.
\code{"+"} means True is better, \code{"-"} means False is better.}

\item{Nmax}{Maximum number of selected taxa.}

\item{alpha}{Regularization parameter when \code{fitf = "regul"}.}

\item{maxIter}{Maximum number of GA iterations.}

\item{maxFitness}{Early stopping threshold for fitness value.}

\item{run}{Number of consecutive generations without improvement before stopping.}

\item{popSize}{Population size for genetic algorithm.}

\item{parallel}{Logical. Whether to enable parallel computing.}

\item{monitor}{Logical or function. If set to "save", fitness evolution will be stored.}
}
\value{
A list containing:
\itemize{
\item fitness: Best fitness value obtained.
\item x: Binary selection vector.
\item members: Names of selected taxa.
\item abundance: Aggregated topological abundance vector.
\item performance: Correlation (continuous y) or eta coefficient (discrete y).
\item fitnesses: Fitness values across generations (if monitored).
\item correlations: Correlations across generations (if monitored).
}
}
\description{
This function uses a genetic algorithm to discover functional co-response group
(columns from matrix \code{M}) that maximizes correlation or association with a functional variable \code{y}.
}
\details{
The optimization problem is solved using the \code{GA::ga()} function with binary encoding.
For continuous response variables, Pearson correlation is used as fitness.
For discrete outcomes, the eta coefficient is computed as fitness.
}
\examples{
\dontrun{
data <- gFlora::load_example_data()
M <- data$M
y <- data$y
out <- discover(M, y, Nmax=5)
}

}
\seealso{
\code{\link[GA]{ga}}
}
