% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_example_data.R
\name{load_example_data}
\alias{load_example_data}
\title{Load example dataset for gFlora}
\usage{
load_example_data()
}
\value{
A list with \code{M} (matrix/data.frame) and \code{y} (vector).
}
\description{
Load example dataset for gFlora
}
