% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover.R
\name{discover}
\alias{discover}
\title{graph convolution-based functional co-response group discovery
https://doi.org/10.1109/TCBBIO.2025.3560853}
\usage{
discover(
  M,
  y = NULL,
  pk = NULL,
  fit_f = "nmax",
  y_type = "continuous",
  fun_s = "+",
  n_max = NULL,
  alpha = 40,
  max_iter = 500,
  max_fitness = 1,
  run = 100,
  pop_size = 200,
  parallel = TRUE,
  monitor = FALSE
)
}
\arguments{
\item{M}{Numeric matrix (samples × taxa). values are topological abundance.}

\item{y}{Functional variable. Can be continuous, binary or categorical depending on \code{y_type}.}

\item{pk}{Optional vector of prior knowledge (same length as ncol(M)).
Elements equal to 1 indicate mandatory inclusion, 0 otherwise.}

\item{fit_f}{Fitness function type. Options: \code{"nmax"} or \code{"regul"}.}

\item{y_type}{Type of functional variable. Options: \code{"continuous"} or \code{"binary"} or \code{"categorical"}.}

\item{fun_s}{Direction of association for binary outcome.
\code{"+"} means True is better, \code{"-"} means False is better.}

\item{n_max}{Maximum number of selected taxa.}

\item{alpha}{Regularization parameter when \code{fit_f = "regul"}.}

\item{max_iter}{Maximum number of GA iterations.}

\item{max_fitness}{Early stopping threshold for fitness value.}

\item{run}{Number of consecutive generations without improvement before stopping.}

\item{pop_size}{Population size for genetic algorithm.}

\item{parallel}{Logical. Whether to enable parallel computing.}

\item{monitor}{Logical or function. If set to "save", fitness evolution will be stored.}
}
\value{
A list containing:
\itemize{
\item fitness: Best fitness value obtained.
\item x: Binary selection vector.
\item members: Names of selected taxa from colomn names of M.
\item abundance: Aggregated topological abundance vector.
\item performance: Correlation (continuous y) or eta coefficient (discrete y).
\item fitnesses: Fitness values across generations (if monitored).
}
}
\description{
This function uses a genetic algorithm to discover functional co-response group
(columns from matrix \code{M}) that maximizes correlation or association with a functional variable \code{y}.
}
\details{
The optimization problem is solved using the \code{GA::ga()} function with binary encoding.
For continuous response variables, Pearson correlation is used as fitness.
For discrete outcomes, the eta coefficient is computed as fitness.
}
\examples{
\dontrun{
data <- gFlora::load_example_data()
M <- data$M
y <- data$y
out <- discover(M, y, n_max=5)
}

}
\seealso{
\code{\link[GA]{ga}}
}
